fixed calculate the transport key

+++ CardCheck/CardCheck/Kernel/Model/KeyChainData/KeyChainData.h
@@ -41,7 +41,7 @@
 /*
  одноразовый пароль / OTP: шестизначное число
  */
-@property (nonatomic) long otp;
+@property (nonatomic) int otp;
 
 
 #pragma mark - Init
